+++ noodles-bam/src/reader/record.rs
@@ -96,6 +96,8 @@ where
 
     get_data(src, record.data_mut())?;
 
+    resolve_cigar(header, record)?;
+
     Ok(())
 }
 
@@ -167,6 +169,48 @@ where
     }
 
     Ok(src.get_i32_le())
+}
+
+// § 4.2.2 "`N_CIGAR_OP` field" (2022-08-22)
+fn resolve_cigar(header: &sam::Header, record: &mut Record) -> io::Result<()> {
+    use sam::record::{cigar::op, data::field::Tag};
+
+    if let Some(id) = record.reference_sequence_id() {
+        if let [op_0, op_1] = record.cigar().as_ref() {
+            if op_0.kind() == op::Kind::SoftClip && op_1.kind() == op::Kind::Skip {
+                let k = record.sequence().len();
+
+                let m = header
+                    .reference_sequences()
+                    .get_index(id)
+                    .map(|(_, rs)| rs.length().get())
+                    .ok_or_else(|| {
+                        io::Error::new(io::ErrorKind::InvalidData, "invalid reference sequence id")
+                    })?;
+
+                if op_0.len() == k && op_1.len() == m {
+                    if let Some((_, value)) = record.data_mut().remove(Tag::Cigar) {
+                        let data = value.as_uint32_array().ok_or_else(|| {
+                            io::Error::new(
+                                io::ErrorKind::InvalidData,
+                                "invalid CG data field value type",
+                            )
+                        })?;
+
+                        let cigar = record.cigar_mut();
+                        cigar.clear();
+
+                        for &n in data {
+                            let op = cigar::decode_op(n)?;
+                            cigar.as_mut().push(op);
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    Ok(())
 }
 
 #[cfg(test)]
@@ -264,6 +308,45 @@ mod tests {
             get_reference_sequence_id(&mut src, 0),
             Err(e) if e.kind() == io::ErrorKind::InvalidData
         ));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_resolve_cigar() -> Result<(), Box<dyn std::error::Error>> {
+        use sam::{
+            header::record::value::{map::ReferenceSequence, Map},
+            record::{
+                cigar::op::{self, Op},
+                data::field::{Tag, Value},
+                Cigar,
+            },
+        };
+
+        let header = sam::Header::builder()
+            .add_reference_sequence(
+                "sq0".parse()?,
+                Map::<ReferenceSequence>::new(NonZeroUsize::try_from(8)?),
+            )
+            .build();
+
+        let mut record = Record::builder()
+            .set_reference_sequence_id(0)
+            .set_cigar("4S8N".parse()?)
+            .set_sequence("ACGT".parse()?)
+            .set_data(
+                [(Tag::Cigar, Value::UInt32Array(vec![0x40]))]
+                    .into_iter()
+                    .collect(),
+            )
+            .build();
+
+        resolve_cigar(&header, &mut record)?;
+
+        let expected = Cigar::try_from(vec![Op::new(op::Kind::Match, 4)])?;
+
+        assert_eq!(record.cigar(), &expected);
+        assert!(record.data().get(Tag::Cigar).is_none());
 
         Ok(())
     }